requirements.txt
discord==2.2.2
requests==2.28.2
twitchAPI==3.9.0
python-dotenv==1.0.0